Fix #1088: refactor operator code - split OperatorUtil and improve code quality
