+++ apps/wanaku-operator/src/main/java/ai/wanaku/operator/util/OperatorConstants.java
@@ -1,0 +1,14 @@
+package ai.wanaku.operator.util;
+
+/**
+ * Shared constants used across operator reconcilers and resource factories.
+ */
+public final class OperatorConstants {
+
+    /**
+     * The capability type identifier for Camel Integration Capability deployments.
+     */
+    public static final String CAMEL_INTEGRATION_CAPABILITY_TYPE = "camel-integration-capability";
+
+    private OperatorConstants() {}
+}
